+++ packages/frontend-misskey-world-engine/vitest.config.ts
@@ -3,5 +3,6 @@ export default {
 		include: ['./test/**/*.test.ts'],
 		environment: 'node',
 		globals: true,
+		execArgv: ['--no-warnings'],
 	},
 };